% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{scale_color_highres}
\alias{scale_color_highres}
\title{Conveniently show color- and shape scale with pretty legend}
\usage{
scale_color_highres(numShapeLevels, colorPalette = scales::hue_pal(), ...)
}
\arguments{
\item{numShapeLevels}{number of distinct shapes to use}

\item{colorPalette}{expects a color scale (such as those provided by scales::hue_pal() or function(x) RColorBrewer::brewer.pal(x, "Set1")),}
}
\description{
This function is a convenience wrapper to allow concurrent color- and shape scaling
including a well-formatted legend that shows both color- and shape information for values
(an approach I often use to distinguish > 6 or so different groups).
}
